# artifyme-backend/docs/swagger.yaml
basePath: /
definitions:
  artists.AddPaintingRequest:
    properties:
      category:
        $ref: '#/definitions/auth.PaintingCategory'
      description:
        type: string
      image_url:
        type: string
      title:
        type: string
    required:
    - category
    - image_url
    - title
    type: object
  artists.ArtistPainting:
    properties:
      artist_id:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      image_url:
        type: string
      painting_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  artists.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: ganesh@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.PaintingCategory:
    enum:
    - PENCIL
    - CHARCOAL
    - OIL
    - ACRYLIC
    - WATERCOLOR
    type: string
    x-enum-varnames:
    - Pencil
    - Charcoal
    - Oil
    - Acrylic
    - WaterColor
  auth.RegisterRequest:
    properties:
      address:
        example: Pune, India
        type: string
      email:
        example: ganesh@example.com
        type: string
      full_name:
        example: Ganesh Terkar
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      phone_number:
        example: +91XXXXXXXXXX
        type: string
      role:
        allOf:
        - $ref: '#/definitions/auth.UserRole'
        example: CUSTOMER
    required:
    - address
    - email
    - full_name
    - password
    - phone_number
    type: object
  auth.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.UserDTO:
    properties:
      address:
        example: Pune, India
        type: string
      created_at:
        type: string
      email:
        example: ganesh@example.com
        type: string
      full_name:
        example: Ganesh Terkar
        type: string
      phone_number:
        example: +91XXXXXXXXXX
        type: string
      role:
        example: CUSTOMER
        type: string
      status:
        example: ACTIVE
        type: string
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      user_name:
        example: ganesh123
        type: string
    type: object
  auth.UserRole:
    enum:
    - ADMIN
    - ARTIST
    - CUSTOMER
    - DELIVERY_AGENT
    type: string
    x-enum-varnames:
    - Admin
    - Artist
    - Customer
    - DeliveryAgent
host: localhost:8080
info:
  contact:
    email: ganesh@example.com
    name: Ganesh Terkar
  description: Backend API for ArtifyMe platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ArtifyMe Backend API
  version: "1.0"
paths:
  /api/artist/paintings/:
    get:
      description: Get all paintings added by the logged-in artist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/artists.ArtistPainting'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/artists.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/artists.ErrorResponse'
      summary: Get my paintings
      tags:
      - Artist
  /api/artist/paintings/add:
    post:
      consumes:
      - application/json
      description: Artist adds a painting
      parameters:
      - description: Painting details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/artists.AddPaintingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/artists.ArtistPainting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/artists.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/artists.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/artists.ErrorResponse'
      summary: Add a new painting
      tags:
      - Artist
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user using email and password and return JWT token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Token generation failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account. Defaults role to CUSTOMER if not provided.
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"